refactor: extract url-to-file-info parser
