+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/legacy/download/utility/MiniAppFileWriterTest.kt
@@ -31,7 +31,7 @@ class MiniAppFileWriterTest : DownloadBaseTest() {
 
     @Test
     fun shouldInjectUrlParser() {
-        assertThat(miniWriter.localUrlParser).isNotNull()
+        assertThat(miniWriter.urlToFileInfoParser).isNotNull()
     }
 
     @Test
